Altering error handling and debug messages

+++ CommunicationLayer/CommDeviceControl/UdpMessageForwarder.h
@@ -46,5 +46,5 @@ private:
     AmqpClient::Channel::ptr_t channel_;
     QString exchangeName_;
     QString ipAddress_;
-    quint64 udpPort_;
+    quint16 udpPort_;
 };
